rename recompute_cpu_offload && recover simple_cpu_offload
